api/requirements.txt
django==5.1.8
djangorestframework==3.16.0
django-cors-headers==4.7.0
anthropic==0.49.0
python-dotenv==1.1.0
rich==13.7.0
requests==2.32.3
PyPDF2
python-docx 